Create directory for random tools
Create a misc/ directory to collect small random tools. They are tools
that only need a single file and, therefore, a entire directory for each
one of them would be an overkill.

Populate this new directory misc/ with a first tool and a README file.

The README file should explain briefly what each tool in this directory
does.

The first tool being added is txt2pdf.py